Move generated models to `models`-package
